pyproject.toml
[project]
name = "bruggeman"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Davíd Brakenhoff", email = "d.brakenhoff@artesia-water.nl" },
]
requires-python = ">=3.12"
dependencies = ["numpy", "scipy"]
dynamic = ["version"]

[project.scripts]
bruggeman = "bruggeman:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "bruggeman/__about__.py"

[tool.hatch.metadata]
allow-direct-references = true

[project.optional-dependencies]
# NOTE: it is recommended to install the latest version of latexify-py. Unfortunately
# PyPI won't accept this as a dependency. So we use the official release, but for the
# best results use the following in pyproject.toml:
#  - "latexify-py@git+https://github.com/google/latexify_py.git#egg=main"
# or install using:
#  - pip install git+https://github.com/google/latexify_py.git#egg=main
latexify = ["latexify-py"]
ci = ["pytest", "pytest-cov", "nbformat", "nbconvert", "mypy", "ruff"]
docs = [
    "bruggeman[latexify]",
    "Ipython",
    "ipykernel",
    "pydata-sphinx-theme",
    "sphinx",
    "sphinxcontrib-bibtex",
    "numpydoc",
    "sphinx-design",
    "myst_nb",
    "ttim",
]

[tool.ruff]
extend-include = ["*.ipynb"]
lint.extend-select = ["I"]
show-fixes = true
fix = true

[tool.pytest.ini_options]
addopts = "--strict-markers --durations=0"
markers = ["notebooks: run notebooks"]

[tool.mypy]
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["scipy.special.*", "latexify.*"]
ignore_missing_imports = true
